chop_css

Function for separating the conditioned responses into multiple pieces for two CSs.

chop_css(cs1, cs2, data, subj, cs_paired = NULL, group = NULL, na.rm = FALSE)

Arguments

cs1

The column name(s) of the conditioned responses for the first conditioned stimulus

cs2

The column name(s) of the conditioned responses for the second conditioned stimulus

data

A data frame containing all the relevant columns for the analyses

subj

The name of the column including the participant numbers. Unique numbers are expected

cs_paired

A character vector with the trials that were paired. Default is set to NULL, suggesting that there was full reinforcement @return A tibble with the following column names -- separately forr cs1 and cs2: "id" = id column; "top10per" = mean of top 10 "bottom10per" = mean of bottom 10 "minfirst = all trials minus the first one; "all" = mean of all trials; "t33per = mean of top 33 "m33per" = mean of medium 33 "b33per" = mean of bottom 33 "fhalf" = mean of first half of the trials; "lhalf" = mean of last half of the trials; "ftrial" = first trial; "ltrial" = last trial; "t20per" = mean of top 20 "b20per" = mean of bottom 20 "f2trial" = mean of first two trials; "l2trial" = mean of last two trials; "per2_X" = mean per two trials (i.e., block), with X referring to the number of block.

group

the name of the group, if included, default to NULL

na.rm

Whether NAs should be removed, default to FALSE