forestplot_mf

Basic function for forest plot

forestplot_mf(data, ci = TRUE, include_label_text = TRUE, ...)

Arguments

data
a universe_mf or multiverse_mf object

ci
should confidence intervals be included -- default to TRUE

include_label_text
Whether the labels for each effect should be include. Default to TRUE

...
any additional argument

Value

A plot #'

Details

This is a wrapper around the forestplot::forestplot function. The function only uses the ANOVAs and the t-tests. For the t-tests though it includes only the two-sided, as they are the same with the one-sided ones and having both would probably give a false picture of the effect.

The include_label_text can be used when multiple plots need to be combined side by side, so one of them does not include the labels.