inference_cs

Function for summarizing the multiverse results.

inference_cs(data, alpha_level = 0.05, na.rm = FALSE, framework = "Both")

Arguments

data

a data frame with the results of a multiverse analyses.

alpha_level

What should be the alpha level used (default to 0.05).

na.rm

Should NA's be removed (default to FALSE). See details.

framework

Inference framework. Values could be "NHST", "Bayesian", or "Both" (no case sensitivity)

Value

A data frame with summaries of the results.

Details

For now the function returns mean, median, standard deviations of p values and proportion of p values below a criterion defined by the alpha_level parameter (default to 0.05) as well as mean Bayes factors (please see the `framework` argument). The user may choose to drop the NAs for the summary statistic.