lm_mf

Basic function for running frequentist's regression models

lm_mf(
  cs1,
  cs2,
  data,
  subj,
  group = NULL,
  na.rm = FALSE,
  paired = TRUE,
  phase = "acquisition",
  dv = "scr",
  exclusion = "full data"
)

Arguments

cs1
The column name(s) of the conditioned responses for the first conditioned stimulus.

cs2
The column name(s) of the conditioned responses for the second conditioned stimulus.

data
A data frame containing with all the relevant columns

subj
The name of the column including the participant numbers. Unique numbers are expected.

group
the name of the group, if included, default to NULL.

na.rm
Whether NAs should be removed, default to FALSE.

paired
Whether the t-test refers to a paired or independent sample(s). Default to TRUE.

phase
The conditioned phase that the analyses refer to. Accepted values are acquisition, or acq, extinction, or ext.

dv
name of the measured conditioned response. Default to "SCR".

exclusion
If any exclusion criteria were applied, default to full data

Value

A tibble with the following column names: x: the name of the independent variable (e.g., cs) y: the name of the dependent variable as this defined in the dv argument exclusion: see exclusion argument model: the model that was run (e.g., t-test) controls: ignore this column for this test method: the model that was run p.value: the p-value of the test effect.size: the estimated effect size estimate: the estimate of the test run. For the t-test is the mean of the differences statistic: the t-value conf.low: the lower confidence interval for the estimate conf.high: the higher confidence interval for the estimate framework: were the data analysed within a NHST or Bayesian framework? data_used: a list with the data used for the specific test

Details

Given the correct names for the cs1, cs2, subj, and data, the function will run one- and two-sided frequentist's t-tests. In case cs1 or cs2 refer to multiple columns, the mean -- per row -- for each one of these variables will be computed first before running the test. Please note that cs1 is implicitly referred to the cs that is reinforced, and cs2 to the cs that is not reinforced. Depending on whether the data refer to an acquisition or extinction phase (as defined in the phase argument), the function will return a positive one sided, or negative one-sided, t-test in addition to the two-sided t-test. The returned effect size is Hedge's g.

The function by default runs a Welch t-test, meaning it assumes unequal variances. This is due to calls that such a test should be preferred over Student t-test, at least for paired samples t-test. Please note that if we let R decide which test to run -- this is done by default in stats::t.test, then for some test there would be a Student t-test whereas in some others not.

Examples

# Load example data
data(example_data)

lm_mf(cs1 = "CSP1", cs2 = "CSM1", subj = "id", data = example_data)

#> # A tibble: 1 x 14
#>   x     y     exclusion model controls method p.value effect.size estimate
#>   <chr> <chr> <chr>     <chr> <lgl>    <chr>    <dbl>       <dbl>    <dbl>
#> 1 cs    scr   full data lm    NA       lm      0.0208       0.523    0.523
#> # … with 5 more variables: statistic <dbl>, conf.low <lgl>, conf.high <lgl>,
#> #   framework <chr>, data_used <list>