multiverse_cs

Basic function for conducting multiverse analyses of conditioning data

multiverse_cs(
  cs1,
  cs2,
  data,
  subj,
  group = NULL,
  cs_paired = NULL,
  include_bayes = TRUE,
  include_mixed = FALSE,
  phase = "acquisition",
  cutoff = c(0, 1, 2, 3),
  print_output = TRUE,
  correction = FALSE,
  meta.effect = "d_to_eta2"
)

Arguments

cs1

The column name(s) of the conditioned responses for the first conditioned stimulus

cs2

The column name(s) of the conditioned responses for the second conditioned stimulus

data

A data frame containing all the relevant columns for the analyses

subj

The name of the column including the participant numbers. Unique numbers are expected

group

the name of the group, if included, default to NULL

cs_paired

A character vector with the trials that were paired. Default is set to NULL, suggesting that there was full reinforcement

include_bayes

Whether the Bayesian analyses should be run. Default to TRUE

include_mixed

Whether the mixed model results should be run. Default to FALSE

phase

The conditioned phase that the analyses refer to. Accepted values are acquisition, acq, extinction, or ext

cutoff

A numeric vector of the cutoff criteria applied. Default to 0, 0.05, .1

print_output

Whether to print the output or not. Default set to TRUE

correction

whether the Greenhouse-Geisser correction should be applied or not. Default to FALSE

meta.effect

How the meta-analytic effect should be computed.

Value

A tibble with the following column names: x: the name of the independent variable (e.g., cs) y: the name of the dependent variable as this defined in the dv argument exclusion: see exclusion argument model: the model that was run (e.g., t-test) controls: ignore this column for this test method: the method used p.value: the reported p-value effect.size: the reported effect size estimate: the estimate of the test run statistic: the value of the test conf.low: the lower confidence interval for the estimate conf.high: the higher confidence interval for the estimate framework: were the data analysed within a NHST or Bayesian framework? data_used: a list with the data used for the specific test

Details

In case of higher order interaction, only the highest order effect is returned.

In case the CSs include only 1 observation per participant, or of unequal numbers of CS trials, the function will return the warning ""Skipping ANOVA due to the number of trials for the cs1 and/or cs2."".

In principle the multiverse_cs function runs the universe_cs function multiple times, so whatever holds for the universe_cs -- e.g., in terms of warnings, holds for here as well.