mixed_mf

Basic function for running mixed models for the multiverse analysis

mixed_mf(
  cs1,
  cs2,
  data,
  subj,
  group = NULL,
  phase = "acquisition",
  dv = "scr",
  exclusion = "full data"
)

Arguments

cs1
The column name(s) of the conditioned responses for the first conditioned stimulus

cs2
The column name(s) of the conditioned responses for the second conditioned stimulus

data
A data frame containing with all the relevant columns

subj
The name of the column including the participant numbers. Unique numbers are expected.

group
the name of the group, if included, default to NULL.

phase
The conditioned phase that the analyses refer to. Accepted values are acquisition, or acq, extinction, or ext.

dv
name of the measured conditioned response. Default to "SCR".

exclusion
If any exclusion criteria were applied, default to full data

Value

A data frame with the results.

The data frame returned is the standard one returned in all function in the package. Specifically we have: A tibble with the following column names: x: the name of the independent variable (e.g., cs). There, you can see the term of the model that is returned. So, not the full model is returned but only this particular term. y: the name of the dependent variable as this defined in the dv argument exclusion: see exclusion argument model: the model that was run (e.g., t-test) controls: ignore this column for this test method: the model that was run p.value: irrelevant here effect.size: irrelevant here estimate: the estimate of the test run statistic: the t-value conf.low: the lower confidence interval for the estimate conf.high: the higher confidence interval for the estimate data_used: a list with the data used for the specific test

Details

The function assumes that you include more than 1 trial per CS. The function returns an error if that is not the function.

The function performs by default two dependent variable standardizations, the one per subject and the other one without taking subject into account.

In case time is included, the function computes the intercept -- i.e., the 0 point -- on the middle of the time sequence.

The following models are run and compared: a) Intercept only model, b) Intercept plus CS model, and c) Intercept plus CS x Time interaction.

Separate models are run with `Subject` as random factor, as well as `Subject and Time` as random factors.

The model is fit by maximizing the log-likelihood (i.e., "ML" term in nlme::lme).

The model comparison is done using `BIC`.

See also