multiverse_cs

Basic function for conducting multiverse analyses of conditioning data

multiverse_cs(
  cs1,
  cs2,
  data,
  subj,
  group = NULL,
  cs_paired = NULL,
  include_bayes = TRUE,
  phase = "acquisition",
  cutoff = c(0, 1, 2, 3),
  print_output = TRUE,
  correction = FALSE
)

Arguments

cs1
The column name(s) of the conditioned responses for the first conditioned stimulus

cs2
The column name(s) of the conditioned responses for the second conditioned stimulus

data
A data frame containing with all the relevant columns

subj
The name of the column including the participant numbers. Unique numbers are expected.

group
the name of the group, if included, default to NULL.

cs_paired
A character vector with the trials that were paired. Default is set to NULL, suggesting that there was full reinforcement

include_bayes
Whether the bayesian analyses should be run. Default to TRUE

phase
The conditioned phase that the analyses refer to. Accepted values are acquisition, or acq, extinction, or ext.

cutoff
A numeric vector of the cutoff criteria applied. Default to 0, 0.05, .1

print_output
Whether to print the output or not. Default set to TRUE

correction
whether the Greenhouse-Geisser correction should be applied or not. Default to FALSE

Value

A tibble with the following column names: x: the name of the independent variable (e.g., cs) y: the name of the dependent variable as this defined in the dv argument exclusion: see exclusion argument model: the model that was run (e.g., t-test) controls: ignore this column for this test method: the model that was run p.value: irrelevant here effect.size: irrelevant here estimate: the estimate of the test run statistic: the t-value conf.low: the lower confidence interval for the estimate conf.high: the higher confidence interval for the estimate data_used: a list with the data used for the specific test.

#' # Load data and define argument to be fed in universe_cs data("example_data", package = "multifear") example_data <- example_data[1:10, ] cs1 <- paste0("CSP", 1:10) cs2 <- paste0("CSM", 1:10) subj = "id" multiverse_cs(cs1, cs2, subj = subj, data = example_data)

Details

In case of higher order interaction, only the highest order effect is returned.