rm_anova_mf

Basic function for running the frequentist's repeated measures ANOVA

rm_anova_mf(
  cs1,
  cs2,
  data,
  subj,
  time = TRUE,
  group = NULL,
  phase = "acquisition",
  dv = "scr",
  exclusion = "full data"
)

Arguments

cs1
The column name(s) of the conditioned responses for the first conditioned stimulus

cs2
The column name(s) of the conditioned responses for the second conditioned stimulus

data
A data frame containing with all the relevant columns

subj
The name of the column including the participant numbers. Unique numbers are expected.

time
should time be included? Default to TRUE.

group
the name of the group, if included, default to NULL.

phase
The conditioned phase that the analyses refer to. Accepted values are acquisition, or acq, extinction, or ext.

dv
name of the measured conditioned response. Default to "SCR".

exclusion
If any exclusion criteria were applied, default to full data

Value

A basic function for running repeated measures ANOVAs.

A tibble with the following column names: x: the name of the independent variable (e.g., cs) y: the name of the dependent variable as this defined in the dv argument exclusion: see exclusion argument model: the model that was run (e.g., t-test) controls: ignore this column for this test method: the model that was run p.value: the p-value of the test effect.size: the estimated effect size estimate: the estimate of the test run statistic: the t-value conf.low: the lower confidence interval for the estimate conf.high: the higher confidence interval for the estimate framework: were the data analysed within a NHST or Bayesian framework? data_used: a list with the data used for the specific test

Details

In case the time argument is set to TRUE (default value), the function will include this as a within subjects factor, assuming that the columns in cs1 and cs2 correspond to ascending time points (e.g., cs1 trial 1, cs1 trial 2 ... cs1 trial n). If this is not the case, the results are not to be trusted.

The function uses the ez::ezANOVA function. The function gives by default a warning regarding the collapsing of factors. This function here suppresses this warning but the user should be aware of it. Please note that at the moment no sphericity correction is performed. The reported effect size is omega squared as this is computed by sjstats::omega_sq.

Examples

# Load example data
data(example_data)

# Briefly define argument values that will be plugged in later on in the functions
cs1 <- paste0("CSP", 1:10)
cs2 <- paste0("CSM", 1:10)
subj <- "id"
group <- "group"

# Repeated measures ANOVA
rm_anova_mf(cs1 = cs1, cs2 = cs2, subj = subj, data = example_data, time = TRUE)

#> Registered S3 methods overwritten by 'lme4':
#>   method                          from
#>   cooks.distance.influence.merMod car 
#>   influence.merMod                car 
#>   dfbeta.influence.merMod         car 
#>   dfbetas.influence.merMod        car 

#> Warning: Currently only supports partial omega squared for repeated-measures ANOVAs.

#> Warning: `...` is not empty.
#> 
#> We detected these problematic arguments:
#> * `needs_dots`
#> 
#> These dots only exist to allow future extensions and should be empty.
#> Did you misspecify an argument?

#> # A tibble: 1 x 14
#>   x     y     exclusion model controls method p.value effect.size estimate
#>   <chr> <chr> <chr>     <chr> <lgl>    <chr>    <dbl>       <dbl> <lgl>   
#> 1 cs:t… scr   full data rep … NA       rep A… 3.70e-9          NA NA      
#> # … with 5 more variables: statistic <dbl>, conf.low <lgl>, conf.high <lgl>,
#> #   framework <chr>, data_used <list>