rm_banova_mf

Basic function for running the Bayesian repeated measures ANOVA

rm_banova_mf(
  cs1,
  cs2,
  data,
  subj,
  time = TRUE,
  group = NULL,
  phase = "acquisition",
  dv = "scr",
  exclusion = "full data",
  multicore = TRUE
)

Arguments

cs1
The column name(s) of the conditioned responses for the first conditioned stimulus

cs2
The column name(s) of the conditioned responses for the second conditioned stimulus

data
A data frame containing with all the relevant columns

subj
The name of the column including the participant numbers. Unique numbers are expected.

time
should time be included? Default to TRUE.

group
the name of the group, if included, default to NULL.

phase
The conditioned phase that the analyses refer to. Accepted values are acquisition, or acq, extinction, or ext.

dv
name of the measured conditioned response. Default to "SCR".

exclusion
If any exclusion criteria were applied, default to full data

multicore
The option to run the analysis in multiple cores, not available under Windows. Default to TRUE.

Value

A tibble with the following column names: x: the name of the independent variable (e.g., cs) y: the name of the dependent variable as this defined in the dv argument exclusion: see exclusion argument model: the model that was run (e.g., t-test) controls: ignore this column for this test method: the model that was run p.value: irrelevant here effect.size: irrelevant here estimate: the estimate of the test run statistic: the t-value conf.low: the lower confidence interval for the estimate conf.high: the higher confidence interval for the estimate framework: were the data analysed within a NHST or Bayesian framework? data_used: a list with the data used for the specific test

Details

In case the time argument is set to true, the function will include this as a within subjects factor, assuming that the columns in cs1 and cs2 corrrespond to ascending time points (e.g., cs1 trial 1, cs1 trial 2 ... cs1 trial n). If this is not the case, the results are not to be trusted.

The ANOVA will run *all* possible models and combinations. Please note that in case of many factors, this will mean that the analysis will take a long time to be completed.